ServerYou/requirements.txt
grpcio
googleapis-common-protos
flask
redis